fix(storage): add html impl. of app_path_provider
