+++ packages/storage/amplify_storage_s3/lib/src/amplify_storage_s3_impl.dart
@@ -15,7 +15,7 @@ import 'dart:async';
 
 import 'package:amplify_core/amplify_core.dart';
 import 'package:amplify_db_common/amplify_db_common.dart' as db_common;
-import 'package:amplify_storage_s3/src/utils/app_path_provider.dart';
+import 'package:amplify_storage_s3/src/utils/app_path_provider/app_path_provider.dart';
 import 'package:amplify_storage_s3_dart/amplify_storage_s3_dart.dart';
 
 /// {@template amplify_storage_s3_.amplify_storage_s3_plugin}
