+++ packages/storage/amplify_storage_s3_dart/lib/src/utils/app_path_provider/app_path_provider.dart
@@ -1,0 +1,17 @@
+// Copyright 2022 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+export 'app_path_provider_stub.dart'
+    if (dart.library.html) 'app_path_provider_html.dart'
+    if (dart.library.io) 'app_path_provider_io.dart';
